Title:feat Fuzzy elements using new settings
Description:模糊图元使用新的设置方式

RootCause:原设置方式有些复杂

Solution:使用新的方式统一设定

+++ service/cmanagerattributeservice.cpp
@@ -130,7 +130,7 @@ void CManagerAttributeService::showSelectedCommonProperty(CDrawScene *scence, QL
             qDebug() << "Text Item = " << propertys;
             break;
         case BlurType://模糊
-            propertys[Blurtype] = static_cast<CGraphicsMasicoItem *>(items.at(0))->getBlurEffect();
+            propertys[Blurtype] = static_cast<int>(static_cast<CGraphicsMasicoItem *>(items.at(0))->getBlurEffect());
             propertys[BlurWith] = static_cast<CGraphicsMasicoItem *>(items.at(0))->getBlurWidth();
             break;
         default:
@@ -383,7 +383,7 @@ void CManagerAttributeService::showSelectedCommonProperty(CDrawScene *scence, QL
             break;
         case BlurType://模糊
             if (propertys.contains(Blurtype)) {
-                if (propertys[Blurtype] == static_cast<CGraphicsMasicoItem *>(item)->getBlurEffect()) {
+                if (propertys[Blurtype] == static_cast<int>(static_cast<CGraphicsMasicoItem *>(item)->getBlurEffect())) {
                     allPropertys[Blurtype] = propertys[Blurtype];
                 } else {
                     allPropertys[Blurtype] = tmpVariant;
